requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
fastar==0.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.48.0
shellingham==1.5.4
six==1.17.0
starlette==0.50.0
typer==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
uvloop==0.22.1
vonage==4.7.2
vonage-account==1.1.1
vonage-application==2.0.1
vonage-http-client==1.5.1
vonage-jwt==1.1.5
vonage-messages==1.6.2
vonage-network-auth==1.0.2
vonage-network-number-verification==1.0.2
vonage-network-sim-swap==1.1.2
vonage-number-insight==1.0.7
vonage-numbers==1.0.5
vonage-sms==1.1.6
vonage-subaccounts==1.0.4
vonage-users==1.2.1
vonage-utils==1.1.4
vonage-verify==2.1.0
vonage-verify-legacy==1.0.1
vonage-video==1.5.1
vonage-voice==1.4.0
watchfiles==1.1.1
websockets==15.0.1
